Fix 3DNow! output
The PFRSQRT instruction only computes the (reciprocal)
square root of the lower element but puts it in both
elements of the destination vector. To compute the
square root the reciprocal was vector-multiplied with
the original value, leading to one correct value and
a distorted version of it.

We use instead standard float instructions. It's still
about 2x faster, so we are doing well.

Also changes the logic slightly in the "Increment k"
section, but that should have no effect on output.